scripts/requirements.txt
requests
pandas
google-cloud-storage
python-dotenv
pyarrow
pytest
requests-mock